changed status bar color handling and remove unused job info bar item
