+++ src/extension.ts
@@ -131,7 +131,6 @@ export async function activate(context: ExtensionContext): Promise<CodeForIBMi> 
       commands.executeCommand("code-for-ibmi.refreshLibraryListView");
       commands.executeCommand("code-for-ibmi.refreshIFSBrowser");
       commands.executeCommand("code-for-ibmi.environment.refresh");
-      commands.executeCommand("code-for-ibmi.updateJobInfoBar");
     });
 
   getPassword = async (connection, prompt) => {
